@@ -1,113 +1,60 @@
 package org.dante.springsecurity.config;
 
-import org.dante.springsecurity.security.AuthFilter;
-import org.dante.springsecurity.service.AuthService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.ProviderManager;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.BeanIds;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
+import org.springframework.security.config.Customizer;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.SavedRequestAwareAuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 
 /**
  * Spring Security5 必须要指定一个 PasswordEncoder
  *
  * @author dante
- * @EnableMethodSecurity(prePostEnabled = true) 开启方法调用鉴权拦截
  */
 @Configuration
-@EnableWebSecurity
 @EnableMethodSecurity(prePostEnabled = true)
 public class SecurityConfig {
 
-    private final AuthService authService;
-
-    public SecurityConfig(AuthService authService) {
-        this.authService = authService;
-    }
-
     @Bean
-    public BCryptPasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public AuthenticationProvider authenticationProvider() {
-        DaoAuthenticationProvider authenticationProvider = new DaoAuthenticationProvider();
-        authenticationProvider.setUserDetailsService(authService);
-        authenticationProvider.setPasswordEncoder(passwordEncoder());
-        return authenticationProvider;
-    }
-
-    @Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration configuration) throws Exception {
-        return configuration.getAuthenticationManager();
-    }
-
-    @Bean
-    public SavedRequestAwareAuthenticationSuccessHandler authenticationSuccessHandler() {
-        SavedRequestAwareAuthenticationSuccessHandler handler = new SavedRequestAwareAuthenticationSuccessHandler();
-        handler.setDefaultTargetUrl("/home");
-        handler.setAlwaysUseDefaultTargetUrl(true);
-        return handler;
-    }
-
-    @Bean
-    public AuthFilter authFilter(AuthenticationManager authenticationManager) {
-        AuthFilter authFilter = new AuthFilter();
-        authFilter.setAuthenticationManager(authenticationManager);
-        authFilter.setAuthenticationSuccessHandler(authenticationSuccessHandler());
-        return authFilter;
-    }
-
-    @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http, AuthFilter authFilter) throws Exception {
-        /*
-        AuthFilter authFilter = new AuthFilter();
-        authFilter.setAuthenticationManager(authenticationManager);
-        authFilter.setAuthenticationSuccessHandler(authenticationSuccessHandler());
-        */
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         http
-			.csrf().disable()
-			.authorizeRequests()
-				.antMatchers("/favicon.ico", "/home").permitAll()
-				.antMatchers("/*").authenticated()
-			.and()
-                .formLogin()
-			.and()
-                .addFilterAt(authFilter, UsernamePasswordAuthenticationFilter.class);
+			.authorizeHttpRequests(auth -> auth
+				.requestMatchers("/favicon.ico", "/home").permitAll()
+				.anyRequest().authenticated()
+            )
+            .formLogin(form -> form
+                .loginPage("/login")    // 指定登录页
+                .defaultSuccessUrl("/")
+                .failureUrl("/login?error=true")
+                .permitAll()
+            )
+            .csrf(Customizer.withDefaults()); // 启用CSRF保护
         return http.build();
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
     }
 
     /**
-    //    Error creating bean with name 'securityConfig': Requested bean is currently in creation: Is there an unresolvable circular reference?
-    //    SecurityConfig（或AuthFilter）在实例化时，某些依赖（比如 AuthenticationManager）没有准备好，导致循环引用。
+     * 身份验证实现
+     */
     @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity http, AuthenticationManager authenticationManager) throws Exception {
-        AuthFilter authFilter = new AuthFilter();
-        authFilter.setAuthenticationManager(authenticationManager);
-        authFilter.setAuthenticationSuccessHandler(authenticationSuccessHandler());
-        http
-                .csrf().disable()
-                .authorizeRequests()
-                .antMatchers("/favicon.ico", "/home").permitAll()
-                .antMatchers("/*").authenticated()
-                .and()
-                .formLogin()
-                .and()
-                .addFilterAt(authFilter, UsernamePasswordAuthenticationFilter.class);
-        return http.build();
-    }*/
+    public AuthenticationManager authenticationManager(UserDetailsService userService, PasswordEncoder passwordEncoder) {
+        DaoAuthenticationProvider daoAuthenticationProvider = new DaoAuthenticationProvider();
+        daoAuthenticationProvider.setUserDetailsService(userService);
+        daoAuthenticationProvider.setPasswordEncoder(passwordEncoder);
+
+        return new ProviderManager(daoAuthenticationProvider);
+    }
+
+
 }
