@@ -6,14 +6,18 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Stream;
 
+import jakarta.annotation.PostConstruct;
 import org.dante.springsecurity.vo.UserVO;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.parameters.P;
+import org.springframework.stereotype.Component;
 
 @Slf4j
+@Component
 public class InitDataConfig {
 
-	private static final List<UserVO> userVos;
+	private static List<UserVO> userVos;
 	
 	public final static String PUBLIC_ACCESS = "PUBLIC_ACCESS";
 	
@@ -23,21 +27,22 @@ public class InitDataConfig {
 	public final static String AUTH_MENU_ADD = "AUTH_MENU_ADD";
 	public final static String AUTH_MENU_DEL = "AUTH_MENU_DEL";
 
-	/**
+	/*
 	 * 初始化用户信息
 	 */
-	static {
+	@PostConstruct
+	public void init() {
 		log.info("初始化用户信息......");
 		userVos = Arrays.asList(
-			new UserVO("dante", "$2a$10$yKyh2RWtar7eUrE9O.67M.NaYA8zpr3CgbBmm7L9V4G/7Ttx.hCf2", "super@163.com",
-				Arrays.asList("足球", "编程", "音乐", "游戏"), 
-				Stream.of(AUTH_USER_ADD, AUTH_USER_DEL, AUTH_MENU_ADD, AUTH_MENU_DEL).collect(toSet())),
-			new UserVO("snake", "$2a$10$yKyh2RWtar7eUrE9O.67M.NaYA8zpr3CgbBmm7L9V4G/7Ttx.hCf2", "kid@163.com",
-					Arrays.asList("足球", "编程", "音乐", "游戏"), 
-					Stream.of(AUTH_USER_ADD, AUTH_USER_DEL).collect(toSet())),
-			new UserVO("youna", "$2a$10$yKyh2RWtar7eUrE9O.67M.NaYA8zpr3CgbBmm7L9V4G/7Ttx.hCf2", "kid@163.com",
-					Arrays.asList("足球", "编程", "音乐", "游戏"), 
-					Stream.of(AUTH_MENU_ADD, AUTH_MENU_DEL).collect(toSet()))
+				new UserVO("dante", "{bcrypt}$2a$10$s0Ta/mltqMGKzSNnpPGBqOyGtNtz4khC/N4r4NBtGwoF7i5xxSOTu", "super@163.com",
+						Arrays.asList("足球", "编程", "音乐", "游戏"),
+						Stream.of(AUTH_USER_ADD, AUTH_USER_DEL, AUTH_MENU_ADD, AUTH_MENU_DEL).collect(toSet())),
+				new UserVO("snake", "{bcrypt}$2a$10$s0Ta/mltqMGKzSNnpPGBqOyGtNtz4khC/N4r4NBtGwoF7i5xxSOTu", "kid@163.com",
+						Arrays.asList("足球", "编程", "音乐", "游戏"),
+						Stream.of(AUTH_USER_ADD, AUTH_USER_DEL).collect(toSet())),
+				new UserVO("youna", "{bcrypt}$2a$10$s0Ta/mltqMGKzSNnpPGBqOyGtNtz4khC/N4r4NBtGwoF7i5xxSOTu", "kid@163.com",
+						Arrays.asList("足球", "编程", "音乐", "游戏"),
+						Stream.of(AUTH_MENU_ADD, AUTH_MENU_DEL).collect(toSet()))
 		);
 	}
 
